+++ packages/integration-sdk/CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to
 
 ### Changed
 
+- Upgrade to `@jupiterone/data-model@0.12.0`
 - `Entity.id` is now `string | string[]` in the `data-model` (see
   [PR](https://github.com/JupiterOne/data-model/pull/44)). Integrations may
   enrich existing entities through mapped relationship `targetEntity.id` values.